Add information about database setup
Amend README.md file, so that it contains information about setting up
the initial state of the database. Currently, it is really simple and
does not require any sophisticated method, but that may change over
time.

Additionaly, format the bit of text about gemset to put an emphasis on
its name.