Add information about ways of deployment
Heroku has some neat features, which allow to setup an integration
between Github and automatically deploy changes made to chosen branch
on-merge.

This has been done, so I've added a note to inform other people that it
had been taken care of and there is nothing in particular to do,
regarding the deployment.