Allow filtering by the difficulty level
As a user, I would like to be able to filter recipes by the difficulty
level, so that I can feel comfortable, while preparing food.

This commit introduces another query params in the /api/recipes endpoint
and allows it to be provided via frontend. The design is really simple
and is a subject of change in near future.