Move styles to the main stylesheet file
For some reason current build process raises following error:
```
Sprockets::DoubleLinkError: Multiple files with the same output path cannot be linked ("application.css")
```
an immediate fix, which suits my needs is to move all of the styles to
the global .css file.

This issue has to be addressed due to the fact that the deployment
process fails on the js build step. Perhaps the solution is to create a
smarter build command, which will be investigated.