Refine text search
Current implementation of the search engine did not allow to sort the
recipes by relevance, as there was no scoring system applied. The
implementation of `SIMILAR TO` function does not provide such features,
so I had to poke around a bit PostgreSQL documentation to see what's
available to use.

Apparently, there is full functionality in place to cover for this
scenario [0]. I've leveraged the usage of `to_tsvector` and `to_tsquery`
to implement both, filtering the data [1] and ordering the results by
the scoring [2].

Given that this text is written on the page
"Since a longer document has a greater chance of containing a query term
it is reasonable to take into account document size, e.g., a
hundred-word document with five instances of a search word is probably
more relevant than a thousand-word document with five instances."
it turns out that the less words in the document match the query, the
higher score it gets, so we need it to be ordered in exactly the
opposite way and that justifies ordering DESC instead of ASC.

The sanitization mechanisms provided by Rails are only available via the
`#where` method, so I leveraged of `ActiveRecord::Base` class to utilise
`ActiveRecord::Sanitization::ClassMethods`[3], which consist of methods
meant to sanitize order clauses. It does not seem to be available via
`#order` method's API [4]. (or at least is not mentioned in there)

Previous implementation contatined one bug, which was this line of code:
```
.joins(:ingredients)
```
Given that the ingredients is a separate table and `#joins` by default
translated to an INNER JOIN, all of the ingredients which matched the
query were returned, but the rest was omitted, thus not rendered in the
frontend part of the application. The fix is trivial and involves
replacing INNER JOIN with an RIGHT OUTER JOIN, so that only the recipies
which match the query are returned, but all of the ingredients are
always returned.

[0] https://www.postgresql.org/docs/current/textsearch-controls.html
[1] https://www.postgresql.org/docs/current/textsearch-controls.html
Section 12.3.3
[2] https://www.postgresql.org/docs/10/functions-textsearch.html Table
9.40
[3] https://api.rubyonrails.org/classes/ActiveRecord/Sanitization/ClassMethods.html#method-i-sanitize_sql_for_order
[4] https://api.rubyonrails.org/classes/ActiveRecord/QueryMethods.html#method-i-order
[5] https://guides.rubyonrails.org/active_record_querying.html#ordering