Remove duplicate entries from the results
It's pretty tricky to do an SELECT DISTINCT  statement on such a complex
query, so I've decided to do a bit of a workaround to address the issue
of RIGHT OUTER JOIN usage, basically, few ingredients may be matched in
one recipe, so for each time it matches SQL returns a result. Normally,
this would be solved via SELECT DISTINCT, but it has to match ORDER
condition and that would look pretty weird, if I've started writing
select statements by hand, just to meet this criteria.

Instead, I've plucked ids of the returned records and let rails do it's
job of providing only one result for and id.