Move eager loading a bit further
The associations do not have to be loaded so early.  There is no need
for them in the filtering & searching services, as those operate on
joins and it only results in more records being loaded into the memory.

It really is only needed in the presentation layer, so I've moved it
there.